backend/main.py
"""
TryOnX Backend API
FastAPI server for AI-powered fashion design and virtual try-on
"""

import os
import json
from fastapi import FastAPI, File, UploadFile, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
from typing import Optional, List
import uvicorn
import firebase_admin
from firebase_admin import credentials, auth
from dotenv import load_dotenv

# Load environment variables from .env
load_dotenv()

# Initialize Firebase Admin SDK
service_account_info = os.getenv("FIREBASE_SERVICE_ACCOUNT")
if service_account_info:
    try:
        # The service account info is a string, so it needs to be parsed as JSON
        cred = credentials.Certificate(json.loads(service_account_info))
        firebase_admin.initialize_app(cred)
    except json.JSONDecodeError:
        raise ValueError("Could not parse FIREBASE_SERVICE_ACCOUNT. Please check the format in your .env file.")
    except Exception as e:
        raise ValueError(f"Error initializing Firebase Admin SDK: {e}")


app = FastAPI(
    title="TryOnX API",
    description="AI-Powered 3D Fashion Designer & Virtual Try-On System",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"], # In production, restrict this to your frontend URL
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# OAuth2 scheme to extract the token from the Authorization header
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# Dependency to verify Firebase ID token
async def get_current_user(token: str = Depends(oauth2_scheme)):
    if not service_account_info:
        # If Firebase is not configured, bypass authentication for development
        return {"uid": "development_user"} 
    try:
        decoded_token = auth.verify_id_token(token)
        return decoded_token
    except auth.InvalidIdTokenError:
        raise HTTPException(status_code=401, detail="Invalid ID token")
    except auth.ExpiredIdTokenError:
        raise HTTPException(status_code=401, detail="Expired ID token")
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Token verification failed: {e}")


# Request Models
class PromptRequest(BaseModel):
    prompt: str
    style: Optional[str] = "fashion"
    negative_prompt: Optional[str] = ""

class SketchRequest(BaseModel):
    image_url: str
    refine_level: Optional[str] = "high"

class FabricRequest(BaseModel):
    image_url: str
    garment_type: Optional[str] = "top"

class ChatMessage(BaseModel):
    message: str
    conversation_id: Optional[str] = None

# Health Check
@app.get("/")
async def root():
    return {"message": "TryOnX API is running", "version": "1.0.0"}

@app.get("/health")
async def health():
    return {"status": "healthy"}


# --- Protected Routes ---

@app.get("/api/user/me")
async def get_user_profile(current_user: dict = Depends(get_current_user)):
    """
    Get the profile of the currently authenticated user.
    """
    uid = current_user.get("uid")
    # You can fetch additional user data from your database here
    return {"status": "success", "user_id": uid}


# AI Prompt-Based Generation
@app.post("/api/generate")
async def generate_design(request: PromptRequest, current_user: dict = Depends(get_current_user)):
    """
    Generate 2D fashion design from text prompt using LLaMA 3 + ControlNet/SDXL
    """
    try:
        # TODO: Integrate with LLaMA 3 for prompt interpretation
        # TODO: Use ControlNet/SDXL for 2D generation
        # For now, return mock response
        return {
            "status": "success",
            "image_url": "https://via.placeholder.com/512x512/6366f1/ffffff?text=Generated+Design",
            "prompt": request.prompt,
            "processing_time": 2.5,
            "user_id": current_user.get("uid")
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Sketch-to-Design
@app.post("/api/sketch/refine")
async def refine_sketch(request: SketchRequest, current_user: dict = Depends(get_current_user)):
    """
    Refine sketch using YOLOv8 for component detection + ControlNet
    """
    try:
        # TODO: Integrate YOLOv8 for component detection
        # TODO: Use ControlNet for refinement
        return {
            "status": "success",
            "refined_image_url": "https://via.placeholder.com/512x512/8b5cf6/ffffff?text=Refined+Design",
            "detected_components": ["sleeve", "collar", "hem"],
            "processing_time": 3.2
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Cloth Image Processing
@app.post("/api/fabric/process")
async def process_fabric(request: FabricRequest, current_user: dict = Depends(get_current_user)):
    """
    Extract fabric texture, color, and folds from cloth image
    """
    try:
        # TODO: Implement fabric texture extraction
        # TODO: Color palette extraction
        # TODO: Fold pattern analysis
        return {
            "status": "success",
            "texture_url": "https://via.placeholder.com/256x256",
            "color_palette": ["#a78bfa", "#818cf8", "#c084fc"],
            "fabric_type": "silk",
            "drape_coefficient": 0.75
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# 2D to 3D Conversion
@app.post("/api/convert/2d-to-3d")
async def convert_to_3d(image_url: str, garment_type: str = "top", current_user: dict = Depends(get_current_user)):
    """
    Convert 2D design to 3D mesh using Blender + Kaolin neural reconstruction
    """
    try:
        # TODO: Integrate Blender Python API
        # TODO: Use Kaolin for neural mesh reconstruction
        return {
            "status": "success",
            "glb_url": "https://example.com/model.glb",
            "fbx_url": "https://example.com/model.fbx",
            "mesh_vertices": 5000,
            "processing_time": 15.3
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# AI Chatbot
@app.post("/api/chat")
async def chat(request: ChatMessage, current_user: dict = Depends(get_current_user)):
    """
    Fashion-specialized LLM assistant powered by LLaMA 3
    """
    try:
        # TODO: Integrate LLaMA 3 via Groq/Replicate API
        # TODO: Implement conversation context management
        return {
            "status": "success",
            "response": "That's a great question! Based on your description, I'd recommend considering the fabric weight and drape.",
            "conversation_id": request.conversation_id or "new_conversation"
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# File Upload
@app.post("/api/upload")
async def upload_file(file: UploadFile = File(...), current_user: dict = Depends(get_current_user)):
    """
    Upload image files (sketches, fabric photos, etc.)
    """
    try:
        # TODO: Save to AWS S3 or local storage
        # TODO: Validate file type and size
        return {
            "status": "success",
            "file_url": f"https://example.com/uploads/{file.filename}",
            "file_type": file.content_type
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    uvicorn.run("main:app", host="127.0.0.1", port=8000, reload=True)
