backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
requests==2.31.0
firebase-admin==6.5.0

# AI/ML Dependencies (optional - install as needed)
# torch==2.1.0
# transformers==4.35.0
# diffusers==0.24.0
# controlnet-aux==0.4.0
# ultralytics==8.0.196  # YOLOv8

# AWS (optional)
# boto3==1.29.0

# LangChain (optional)
# langchain==0.0.350
